opt: 优化 PoolAlloc 和 ThreadAlloc 支持文件和行号， 同时 memdebug 支持 dump ThreadAlloc
git-svn-id: http://192.168.20.11/sequoiadb/trunk@42939 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/utilMemBlockPool.hpp
@@ -112,8 +112,7 @@ namespace engine
     **/
    class _utilMemBlockPool : public SDBObject, public _utilSegmentHandler
    {
-      friend BOOLEAN utilPoolPtrCheck( void *ptr, UINT32 *pUserSize,
-                                       UINT32 *pType ) ;
+      friend BOOLEAN utilPoolPtrCheck( void *ptr, UINT32 *pUserSize, UINT16 *pType ) ;
       /*
          Small: 32,   64,   128
          Mid  : 256,  512,  1024
